refactor(playground): extract remote selection geometry helpers
Move DOM measurement logic into a focused utility and remeasure when the presence user list changes.
