+++ apps/playground/src/modules/lexical-eg-walker/RemoteSelectionLayer.tsx
@@ -1,43 +1,9 @@
-import {
-  type DirectionalSelectionRange,
-  isDirectionalSelectionRange,
-  type PresenceUser,
-} from "@softmaple/awareness";
-import type {
-  LexicalBinding,
-  LogicalSelection,
-  LogicalSelectionPoint,
-  StableBlockSelection,
-} from "@softmaple/binding-lexical";
+import type { PresenceUser } from "@softmaple/awareness";
+import type { LexicalBinding } from "@softmaple/binding-lexical";
 import { type RefObject, useLayoutEffect, useRef, useState } from "react";
+import { measurePeer, type PeerGeometry } from "./remote-selection-geometry";
 
-interface OverlayRect {
-  readonly height: number;
-  readonly left: number;
-  readonly top: number;
-  readonly width: number;
-}
-
-interface PeerGeometry {
-  readonly backward: boolean;
-  readonly caret: OverlayRect | null;
-  readonly peer: PresenceUser;
-  readonly selection: ReadonlyArray<OverlayRect>;
-}
-
-interface DomPoint {
-  readonly node: Node;
-  readonly offset: number;
-}
-
-interface DomUnit {
-  readonly from: number;
-  readonly to: number;
-  readonly node: Node;
-  readonly parent: Node;
-  readonly childIndex: number;
-  readonly type: "text" | "break";
-}
+export { resolvePeerSelection } from "./remote-selection-geometry";
 
 export interface RemoteSelectionLayerProps {
   readonly binding: LexicalBinding | null;
@@ -45,146 +11,6 @@ export interface RemoteSelectionLayerProps {
   readonly selfId: string;
   readonly users: ReadonlyArray<PresenceUser>;
 }
-
-const childIndex = (node: ChildNode): number =>
-  node.parentNode === null
-    ? 0
-    : Array.from(node.parentNode.childNodes).indexOf(node);
-
-const collectUnits = (
-  node: Node,
-  units: DomUnit[],
-  initialOffset = 0,
-): number => {
-  let offset = initialOffset;
-  for (const child of node.childNodes) {
-    if (child.nodeType === Node.TEXT_NODE) {
-      const length = child.textContent?.length ?? 0;
-      units.push({
-        from: offset,
-        to: offset + length,
-        node: child,
-        parent: node,
-        childIndex: childIndex(child),
-        type: "text",
-      });
-      offset += length;
-      continue;
-    }
-    if (!(child instanceof HTMLElement)) continue;
-    if (child.tagName === "BR") {
-      units.push({
-        from: offset,
-        to: offset + 1,
-        node: child,
-        parent: node,
-        childIndex: childIndex(child),
-        type: "break",
-      });
-      offset++;
-      continue;
-    }
-    if (child.tagName === "UL" || child.tagName === "OL") continue;
-    offset = collectUnits(child, units, offset);
-  }
-  return offset;
-};
-
-const resolveDomPoint = (
-  binding: LexicalBinding,
-  point: LogicalSelectionPoint,
-): DomPoint | null => {
-  const key = binding.getBlockIndex().blockIdToNodeKey.get(point.blockId);
-  if (key === undefined) return null;
-  const block = binding.editor.getElementByKey(key);
-  if (block === null) return null;
-  const units: DomUnit[] = [];
-  const length = collectUnits(block, units);
-  const offset = Math.max(0, Math.min(length, point.offset));
-  const text = units.find(
-    (unit) => unit.type === "text" && offset >= unit.from && offset <= unit.to,
-  );
-  if (text !== undefined) {
-    return { node: text.node, offset: offset - text.from };
-  }
-  const next = units.find((unit) => unit.from >= offset);
-  if (next !== undefined) {
-    return { node: next.parent, offset: next.childIndex };
-  }
-  return { node: block, offset: block.childNodes.length };
-};
-
-const collapsedRange = (point: DomPoint): Range => {
-  const range = document.createRange();
-  range.setStart(point.node, point.offset);
-  range.collapse(true);
-  return range;
-};
-
-const relativeRect = (
-  rect: DOMRect,
-  host: DOMRect,
-  minimumWidth = 0,
-): OverlayRect => ({
-  left: rect.left - host.left,
-  top: rect.top - host.top,
-  width: Math.max(minimumWidth, rect.width),
-  height: Math.max(18, rect.height),
-});
-
-const toStableSelection = (
-  selection: DirectionalSelectionRange,
-): StableBlockSelection => selection;
-
-export const resolvePeerSelection = (
-  binding: Pick<LexicalBinding, "resolveSelection">,
-  peer: PresenceUser,
-): LogicalSelection | null => {
-  if (!isDirectionalSelectionRange(peer.selection)) return null;
-  try {
-    return binding.resolveSelection(toStableSelection(peer.selection));
-  } catch {
-    return null;
-  }
-};
-
-const measurePeer = (
-  binding: LexicalBinding,
-  host: HTMLElement,
-  peer: PresenceUser,
-): PeerGeometry | null => {
-  const logical = resolvePeerSelection(binding, peer);
-  if (logical === null) return null;
-  const anchor = resolveDomPoint(binding, logical.anchor);
-  const focus = resolveDomPoint(binding, logical.focus);
-  if (anchor === null || focus === null) return null;
-
-  const anchorRange = collapsedRange(anchor);
-  const focusRange = collapsedRange(focus);
-  const backward =
-    anchorRange.compareBoundaryPoints(Range.START_TO_START, focusRange) > 0;
-  const start = backward ? focus : anchor;
-  const end = backward ? anchor : focus;
-  const selectionRange = document.createRange();
-  selectionRange.setStart(start.node, start.offset);
-  selectionRange.setEnd(end.node, end.offset);
-  const hostRect = host.getBoundingClientRect();
-  const selection = Array.from(selectionRange.getClientRects())
-    .filter((rect) => rect.width > 0 && rect.height > 0)
-    .map((rect) => relativeRect(rect, hostRect));
-  const focusRect =
-    focusRange.getClientRects()[0] ?? focusRange.getBoundingClientRect();
-  const caret =
-    focusRect.height > 0
-      ? relativeRect(focusRect, hostRect, 2)
-      : {
-          left: focusRect.left - hostRect.left,
-          top: focusRect.top - hostRect.top,
-          width: 2,
-          height: 20,
-        };
-  return { backward, caret, peer, selection };
-};
 
 export function RemoteSelectionLayer({
   binding,
@@ -195,11 +21,14 @@ export function RemoteSelectionLayer({
   const [geometry, setGeometry] = useState<ReadonlyArray<PeerGeometry>>([]);
   const usersRef = useRef(users);
   usersRef.current = users;
+  const measureRef = useRef<(() => void) | null>(null);
+  const previousUsersRef = useRef(users);
 
   useLayoutEffect(() => {
     const host = hostRef.current;
     if (binding === null || host === null) {
       setGeometry([]);
+      measureRef.current = null;
       return;
     }
     let frame = 0;
@@ -215,6 +44,7 @@ export function RemoteSelectionLayer({
         );
       });
     };
+    measureRef.current = measure;
     measure();
     const resizeObserver = new ResizeObserver(measure);
     const mutationObserver = new MutationObserver(measure);
@@ -228,6 +58,7 @@ export function RemoteSelectionLayer({
     window.addEventListener("resize", measure);
     window.addEventListener("scroll", measure, true);
     return () => {
+      measureRef.current = null;
       cancelAnimationFrame(frame);
       resizeObserver.disconnect();
       mutationObserver.disconnect();
@@ -236,6 +67,12 @@ export function RemoteSelectionLayer({
       window.removeEventListener("scroll", measure, true);
     };
   }, [binding, hostRef, selfId]);
+
+  useLayoutEffect(() => {
+    if (previousUsersRef.current === users) return;
+    previousUsersRef.current = users;
+    measureRef.current?.();
+  }, [users]);
 
   return (
     <div className="pointer-events-none absolute inset-0 z-10 overflow-hidden">
